+++ mayan/apps/document_states/tests/test_workflow_transition_views.py
@@ -1,10 +1,11 @@
 from mayan.apps.common.tests.base import GenericViewTestCase
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
 
+from ..literals import WIDGET_CLASS_TEXTAREA
 from ..models import WorkflowTransition
 from ..permissions import (
-    permission_workflow_edit, permission_workflow_view,
-    permission_workflow_transition
+    permission_workflow_edit, permission_workflow_transition,
+    permission_workflow_view
 )
 
 from .literals import (
@@ -142,95 +143,6 @@ class WorkflowTransitionViewTestCase(
         )
 
 
-class WorkflowTransitionDocumentViewTestCase(
-    WorkflowTestMixin, WorkflowViewTestMixin, WorkflowTransitionViewTestMixin,
-    GenericDocumentViewTestCase
-):
-    auto_upload_test_document = False
-
-    def setUp(self):
-        super(WorkflowTransitionDocumentViewTestCase, self).setUp()
-        self._create_test_workflow()
-        self.test_workflow.document_types.add(self.test_document_type)
-        self._create_test_workflow_states()
-        self._create_test_workflow_transitions()
-        self._upload_test_document()
-        self.test_workflow_instance = self.test_document.workflows.first()
-
-    def test_workflow_transition_selection_get_view_with_workflow_access(self):
-        self.grant_access(
-            obj=self.test_workflow, permission=permission_workflow_transition
-        )
-        response = self._request_test_workflow_transition_selection_get_view()
-        self.assertEqual(response.status_code, 200)
-
-        self.assertEqual(
-            self.test_workflow_instance.get_current_state(),
-            self.test_workflow_state_1
-        )
-
-    def test_workflow_transition_selection_post_view_with_workflow_access(self):
-        self.grant_access(
-            obj=self.test_workflow, permission=permission_workflow_transition
-        )
-        response = self._request_test_workflow_transition_selection_post_view()
-        self.assertEqual(response.status_code, 302)
-
-        self.assertEqual(
-            self.test_workflow_instance.get_current_state(),
-            self.test_workflow_state_1
-        )
-
-    def test_workflow_transition_execute_view_no_access(self):
-        """
-        Test transitioning a workflow without the transition workflow
-        permission.
-        """
-        response = self._request_test_workflow_transition_execute_view()
-        self.assertEqual(response.status_code, 404)
-
-        # Workflow should remain in the same initial state
-        self.assertEqual(
-            self.test_workflow_instance.get_current_state(),
-            self.test_workflow_state_1
-        )
-
-    def test_workflow_transition_execute_view_with_workflow_access(self):
-        """
-        Test transitioning a workflow by granting the transition workflow
-        permission to the role.
-        """
-        self.grant_access(
-            obj=self.test_workflow, permission=permission_workflow_transition
-        )
-
-        response = self._request_test_workflow_transition_execute_view()
-        self.assertEqual(response.status_code, 302)
-
-        self.assertEqual(
-            self.test_workflow_instance.get_current_state(),
-            self.test_workflow_state_2
-        )
-
-    def test_workflow_transition_execute_view_with_transition_access(self):
-        """
-        Test transitioning a workflow by granting the transition workflow
-        permission to the role.
-        """
-        self.grant_access(
-            obj=self.test_workflow_transition,
-            permission=permission_workflow_transition
-        )
-
-        response = self._request_test_workflow_transition_execute_view()
-        self.assertEqual(response.status_code, 302)
-
-        self.assertEqual(
-            self.test_workflow_instance.get_current_state(),
-            self.test_workflow_state_2
-        )
-
-
 class WorkflowTransitionEventViewTestCase(
     WorkflowTestMixin, WorkflowTransitionEventViewTestMixin,
     GenericDocumentViewTestCase
@@ -257,23 +169,31 @@ class WorkflowTransitionEventViewTestCase(
         self.assertEqual(response.status_code, 200)
 
 
+class WorkflowTransitionFieldTestMixin:
+    def _create_test_workflow_transition_field(self, extra_data=None):
+        kwargs = {
+            'field_type': TEST_WORKFLOW_TRANSITION_FIELD_TYPE,
+            'name': TEST_WORKFLOW_TRANSITION_FIELD_NAME,
+            'label': TEST_WORKFLOW_TRANSITION_FIELD_LABEL,
+            'help_text': TEST_WORKFLOW_TRANSITION_FIELD_HELP_TEXT
+        }
+        kwargs.update(extra_data or {})
+
+        self.test_workflow_transition_field = self.test_workflow_transition.fields.create(
+            **kwargs
+        )
+
+
 class WorkflowTransitionFieldViewTestCase(
-    WorkflowTestMixin, WorkflowTransitionFieldViewTestMixin,
-    WorkflowTransitionViewTestMixin, GenericViewTestCase
+    WorkflowTestMixin, WorkflowTransitionFieldTestMixin,
+    WorkflowTransitionFieldViewTestMixin, WorkflowTransitionViewTestMixin,
+    GenericViewTestCase
 ):
     def setUp(self):
         super(WorkflowTransitionFieldViewTestCase, self).setUp()
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
-
-    def _create_test_workflow_transition_field(self):
-        self.test_workflow_transition_field = self.test_workflow_transition.fields.create(
-            field_type=TEST_WORKFLOW_TRANSITION_FIELD_TYPE,
-            name=TEST_WORKFLOW_TRANSITION_FIELD_NAME,
-            label=TEST_WORKFLOW_TRANSITION_FIELD_LABEL,
-            help_text=TEST_WORKFLOW_TRANSITION_FIELD_HELP_TEXT
-        )
 
     def test_workflow_transition_field_list_view_no_permission(self):
         self._create_test_workflow_transition_field()
@@ -353,4 +273,37 @@ class WorkflowTransitionFieldViewTestCase(
         self.assertEqual(
             self.test_workflow_transition.fields.count(),
             workflow_transition_field_count - 1
+        )
+
+
+class WorkflowInstanceTransitionFieldViewTestCase(
+    WorkflowTestMixin, WorkflowTransitionFieldTestMixin,
+    WorkflowTransitionFieldViewTestMixin, WorkflowTransitionViewTestMixin,
+    GenericDocumentViewTestCase
+):
+    def setUp(self):
+        super().setUp()
+        self._create_test_workflow(add_document_type=True)
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field(
+            extra_data={
+                'widget': WIDGET_CLASS_TEXTAREA
+            }
+        )
+        self._upload_test_document()
+        self.test_workflow_instance = self.test_document.workflows.first()
+
+    def test_workflow_transition_text_area_widget_execute_view_with_transition_access(self):
+        self.grant_access(
+            obj=self.test_workflow_transition,
+            permission=permission_workflow_transition
+        )
+
+        response = self._request_test_workflow_transition_execute_view()
+        self.assertEqual(response.status_code, 302)
+
+        self.assertEqual(
+            self.test_workflow_instance.get_current_state(),
+            self.test_workflow_state_2
         )