+++ HISTORY.rst
@@ -213,6 +213,7 @@
 - Autoimport search.py modules from apps.
 - Make ``SearchField`` label optional. If not specified, the ``verbose_name``
   of the model field will be used instead.
+- Sort search form fields.
 
 3.4.14 (2020-08-18)
 ===================